+++ R/controller.R
@@ -51,8 +51,9 @@ MWController <- setRefClass(
              "returnFunc", "initialized"),
   methods = list(
 
-    initialize = function(expr, inputs, autoUpdate = list(value = TRUE, initBtn = FALSE, showCompare = TRUE), nrow = NULL,
-                          ncol = NULL, returnFunc = function(widget, envs) {widget}) {
+    initialize = function(expr, inputs, autoUpdate = list(value = TRUE, initBtn = FALSE, showCompare = TRUE,
+                                                          saveBtn = TRUE, exportBtn = FALSE, exportType = "html2canvas"),
+                          nrow = NULL, ncol = NULL, returnFunc = function(widget, envs) {widget}) {
       expr <<- expr
       inputList <<- inputs$inputList
       uiSpec <<- inputs
@@ -223,11 +224,13 @@ MWController <- setRefClass(
       res
     },
 
-    getModuleUI = function(gadget = TRUE, saveBtn = TRUE, addBorder = !gadget) {
+    getModuleUI = function(gadget = TRUE, saveBtn = TRUE, exportBtn = FALSE, exportType = "html2canvas", addBorder = !gadget) {
       function(ns, okBtn = gadget, width = "100%", height = "400px", fillPage = TRUE) {
-        #ns <- shiny::NS(id)
+        #ns <- shiny::NS
         mwUI(ns, uiSpec, nrow, ncol, outputFunc,
-             okBtn = okBtn, updateBtn = !autoUpdate$value, saveBtn = saveBtn,
+             okBtn = okBtn, saveBtn = autoUpdate$saveBtn,
+             exportBtn = autoUpdate$exportBtn, exportType = autoUpdate$exportType,
+             updateBtn = !autoUpdate$value,
              areaBtns = length(uiSpec$inputs$ind) > 1, border = addBorder,
              width = width, height = height, fillPage = fillPage,
              showCompare = autoUpdate$showCompare)
@@ -300,8 +303,20 @@ MWController <- setRefClass(
             paste('mpWidget-', Sys.Date(), '.html', sep='')
           },
           content = function(con) {
-            htmlwidgets::saveWidget(widget = onDone(controller, stopApp = FALSE),
+            htmlwidgets::saveWidget(widget = onDone(controller$clone(), stopApp = FALSE),
                                     file = con, selfcontained = TRUE)
+          }
+        )
+
+        output$export <- shiny::downloadHandler(
+          filename = function() {
+            'mp-export.png'
+          },
+          content = function(con) {
+            tmp_html <- tempfile(fileext=".html")
+            htmlwidgets::saveWidget(widget = onDone(controller$clone(), stopApp = FALSE),
+                                    file = tmp_html, selfcontained = TRUE)
+            webshot::webshot(url = tmp_html, file = con)
           }
         )